requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
langchain
deep-translator
google-generativeai
python-docx
pdfplumber
numpy